# Conflicts:
#   IDPS/Detector.cpp
#   IDPS/Detector.h
#   IDPS/IDPS.vcxproj.filters
#   IDPS/Packets/Packet.cpp
#   IDPS/Packets/Packet.h
#   IDPS/Sender.h
#   IDPS/Source.cpp

+++ IDPS/ArpTable.h
@@ -15,7 +15,7 @@ public:
     mac getMac(in_addr ipAddr) const noexcept;
 
 private:
-    std::unordered_map<unsigned long, mac> m_table;
+    std::unordered_map<uint32_t, mac> m_table;
     std::string m_fileName;
     PIP_ADDR_STRING m_localAddress = nullptr;
